+++ src/modules/default/index.js
@@ -5,6 +5,7 @@ import * as mongo from '../mongo/';
 import debug from 'debug';
 import { resolvePath } from '../utils';
 import sh from 'shelljs';
+import fs from 'fs';
 
 const log = debug('mup:module:default');
 
@@ -21,14 +22,40 @@ export function help() {
 export function init() {
   log('exec => mup init');
 
-  // TODO check if mup.js or settings.json files exists
   const mupJs = resolvePath(__dirname, 'template/mup.js.sample');
   const settinsJson = resolvePath(__dirname, 'template/settings.json');
   const mupJsDst = resolvePath(process.cwd(), 'mup.js');
   const settingsJsonDst = resolvePath(process.cwd(), 'settings.json');
+  const mupJsExists = fs.existsSync(mupJsDst);
+  const settingsJsonExist = fs.existsSync(settingsJsonDst);
 
-  sh.cp(mupJs, mupJsDst);
-  sh.cp(settinsJson, settingsJsonDst);
+  if (!settingsJsonExist) {
+    sh.cp(settinsJson, settingsJsonDst);
+    console.log('Created settings.json');
+  } else {
+    console.log('Skipping creation of settings.json.');
+    console.log(`settings.json already exist at ${settingsJsonDst}.`);
+  }
+
+  if (!mupJsExists) {
+    sh.cp(mupJs, mupJsDst);
+
+    console.log('Created mup.js');
+    console.log('');
+    console.log('Next Steps:');
+    console.log('');
+    console.log('  Open mup.js and edit the config to meet your needs.');
+    console.log('  Required changes have been marked with a TODO comment.');
+    console.log('');
+    console.log('  Available options can be found in the docs at');
+    console.log('    https://github.com/zodern/meteor-up');
+    console.log('');
+    console.log('  Then run the command:');
+    console.log('    mup setup');
+  } else {
+    console.log('Skipping creation of mup.js');
+    console.log(`mup.js already exists at ${mupJsDst}`);
+  }
 }
 
 export function logs(api) {
@@ -47,17 +74,28 @@ export function restart(api) {
 }
 
 export function setup(api) {
+  function displayNextSteps() {
+    console.log('');
+    console.log('Next, you should run:');
+    console.log('    mup deploy');
+  }
+
   log('exec => mup setup');
   const config = api.getConfig();
-  return docker.setup(api).then(() => {
-    if (config.mongo) {
-      return Promise.all([
-        meteor.setup(api),
-        mongo.setup(api)
-      ]).then(() => mongo.start(api));
-    }
-    return meteor.setup(api);
-  });
+  return docker
+    .setup(api)
+    .then(meteor.setup.bind(null, api))
+    .then(() => {
+      if (config.mongo) {
+        return mongo.setup(api);
+      }
+    })
+    .then(() => {
+      if (config.mongo) {
+        return mongo.start(api);
+      }
+    })
+    .then(displayNextSteps);
 }
 
 export function start(api) {