+++ src/modules/meteor/assets/templates/start.sh
@@ -7,19 +7,32 @@ BUNDLE_PATH=$APP_PATH/current
 ENV_FILE=$APP_PATH/config/env.list
 PORT=<%= port %>
 BIND=<%= bind %>
+NGINX_PROXY_VERSION=latest
+LETS_ENCRYPT_VERSION=latest
 
 # Remove previous version of the app, if exists
 docker rm -f $APPNAME
 
+# Remove container network if still exists
+docker network disconnect bridge -f $APPNAME
+<% for(var network in docker.networks) { %>
+docker network disconnect <%=  docker.networks[network] %> -f $APPNAME
+
+<% } %>
+
 # Remove frontend container if exists
 docker rm -f $APPNAME-frontend
+docker network disconnect bridge -f $APPNAME-frontend
 echo "Removed $APPNAME-frontend"
+
 
 # Remove let's encrypt containers if exists
 docker rm -f $APPNAME-nginx-letsencrypt
+docker network disconnect bridge -f $APPNAME-nginx-letsencrypt
 echo "Removed $APPNAME-nginx-letsencrypt"
 
 docker rm -f $APPNAME-nginx-proxy
+docker network disconnect bridge -f $APPNAME-nginx-proxy
 echo "Removed $APPNAME-nginx-proxy"
 
 # We don't need to fail the deployment because of a docker hub downtime
@@ -34,7 +47,7 @@ docker run \
   <% if(sslConfig && typeof sslConfig.autogenerate === "object")  { %> \
   --expose=80 \
   <% } else { %> \
-  --publish=$BIND:$PORT:80 \
+  --publish=$BIND:$PORT:<%= docker.imagePort %> \
   <% } %> \
   --volume=$BUNDLE_PATH:/bundle \
   --hostname="$HOSTNAME-$APPNAME" \
@@ -48,6 +61,7 @@ docker run \
     -e "VIRTUAL_HOST=<%= sslConfig.autogenerate.domains %>" \
     -e "LETSENCRYPT_HOST=<%= sslConfig.autogenerate.domains %>" \
     -e "LETSENCRYPT_EMAIL=<%= sslConfig.autogenerate.email %>" \
+    -e "HTTPS_METHOD=noredirect" \
   <% } %> \
   --name=$APPNAME \
   <%= docker.image %>
@@ -68,21 +82,21 @@ EOT
 
     # We don't need to fail the deployment because of a docker hub downtime
     set +e
-    docker pull jrcs/letsencrypt-nginx-proxy-companion:latest
-    docker pull jwilder/nginx-proxy
+    docker pull jrcs/letsencrypt-nginx-proxy-companion:$LETS_ENCRYPT_VERSION
+    docker pull jwilder/nginx-proxy:$NGINX_PROXY_VERSION
     set -e
 
     echo "Pulled autogenerate images"
     docker run -d -p 80:80 -p 443:443 \
       --name $APPNAME-nginx-proxy \
       --restart=always \
-      -e "HTTPS_METHOD=noredirect" \
+      -e "DEFAULT_HOST=<%= sslConfig.autogenerate.domains.split(',')[0] %>" \
       -v /opt/$APPNAME/config/nginx-default.conf:/etc/nginx/conf.d/my_proxy.conf:ro \
       -v /opt/$APPNAME/certs:/etc/nginx/certs:ro \
       -v /opt/$APPNAME/config/vhost.d:/etc/nginx/vhost.d \
       -v /opt/$APPNAME/config/html:/usr/share/nginx/html \
       -v /var/run/docker.sock:/tmp/docker.sock:ro \
-      jwilder/nginx-proxy
+      jwilder/nginx-proxy:$NGINX_PROXY_VERSION
       echo "Ran nginx-proxy"
     sleep 15s
 
@@ -92,7 +106,7 @@ EOT
       --volumes-from $APPNAME-nginx-proxy \
       -v /opt/$APPNAME/certs:/etc/nginx/certs:rw \
       -v /var/run/docker.sock:/var/run/docker.sock:ro \
-      jrcs/letsencrypt-nginx-proxy-companion
+      jrcs/letsencrypt-nginx-proxy-companion:$LETS_ENCRYPT_VERSION
     echo "Ran jrcs/letsencrypt-nginx-proxy-companion"
   <% } else { %>
     # We don't need to fail the deployment because of a docker hub downtime