+++ src/modules/meteor/build.js
@@ -1,11 +1,13 @@
 var spawn = require('child_process').spawn;
 var archiver = require('archiver');
 var fs = require('fs');
-
-import { resolvePath } from '../utils';
 var _ = require('underscore');
+import { resolvePath } from '../utils';
+import debug from 'debug';
 
-function buildApp(appPath, buildOptions) {
+const log = debug('mup:module:meteor');
+
+function buildApp(appPath, buildOptions, verbose) {
   // Check if the folder exists
   try {
     fs.statSync(resolvePath(appPath));
@@ -32,7 +34,7 @@ function buildApp(appPath, buildOptions) {
       }
       resolve();
     };
-    buildMeteorApp(appPath, buildOptions, function(code) {
+    buildMeteorApp(appPath, buildOptions, verbose, function(code) {
       if (code === 0) {
         archiveIt(buildOptions.buildLocation, callback);
         return;
@@ -43,7 +45,7 @@ function buildApp(appPath, buildOptions) {
   });
 }
 
-function buildMeteorApp(appPath, buildOptions, callback) {
+function buildMeteorApp(appPath, buildOptions, verbose, callback) {
   var executable = buildOptions.executable || 'meteor';
   var args = [
     'build',
@@ -86,15 +88,29 @@ function buildMeteorApp(appPath, buildOptions, callback) {
     args = ['/c', 'meteor'].concat(args);
   }
 
-  var options = { cwd: appPath };
+  var options = {
+    cwd: appPath,
+    env: {
+      ...process.env,
+      METEOR_HEADLESS: 1
+    },
+    stdio: verbose ? 'inherit' : 'pipe'
+  };
+
+  log(`Build Path: ${appPath}`);
+  log(`Build Command:  ${executable} ${args.join(' ')}`);
+
   var meteor = spawn(executable, args, options);
 
-  meteor.stdout.pipe(process.stdout, { end: false });
-  meteor.stderr.pipe(process.stderr, { end: false });
+  if (!verbose) {
+    meteor.stdout.pipe(process.stdout, { end: false });
+    meteor.stderr.pipe(process.stderr, { end: false });
+  }
 
   meteor.on('error', e => {
     console.log(options);
     console.log(e);
+    console.log('This error usually happens when meteor is not installed.');
   });
   meteor.on('close', callback);
 }