+++ src/modules/mongo/assets/mongo-stop.sh
@@ -1,3 +1,5 @@
 #!/bin/bash
-
+docker update --restart=no mongodb
+docker exec mongodb mongod --shutdown
+sleep 2
 sudo docker rm -f mongodb