+++ src/modules/utils.js
@@ -6,17 +6,42 @@ import expandTilde from 'expand-tilde';
 import fs from 'fs';
 import path from 'path';
 import { promisify } from 'bluebird';
+import stream from 'stream';
+import readline from 'readline';
 
 const log = debug('mup:utils');
 
+function addStdioHandlers(list) {
+  list._taskQueue = list._taskQueue.map(task => {
+    task.options.onStdout = () => {
+      return data => {
+        process.stdout.write(data);
+      };
+    };
+    task.options.onStderr = () => {
+      return data => {
+        process.stderr.write(data);
+      };
+    };
+    return task;
+  });
+}
+
 export function runTaskList(list, sessions, opts) {
+  if (opts && opts.verbose) {
+    addStdioHandlers(list);
+    delete opts.verbose;
+  }
   return new Promise((resolve, reject) => {
     list.run(sessions, opts, summaryMap => {
       for (var host in summaryMap) {
         if (summaryMap.hasOwnProperty(host)) {
           const summary = summaryMap[host];
           if (summary.error) {
-            reject(summary.error);
+            let error = summary.error;
+            error.nodemiralHistory = summary.history;
+            reject(error);
+
             return;
           }
         }
@@ -27,18 +52,55 @@ export function runTaskList(list, sessions, opts) {
   });
 }
 
+// Implments a simple readable stream to pass
+// the logs from nodemiral to readline which
+// then splits it into individual lines.
+class Callback2Stream extends stream.Readable {
+  constructor(options) {
+    // Calls the stream.Readable(options) constructor
+    super(options);
+
+    this.data = [];
+  }
+  addData(data) {
+    if (this.reading) {
+      this.reading = this.push(data);
+    } else {
+      this.data.push(data);
+    }
+  }
+  _read() {
+    this.reading = true;
+    this.data.forEach(() => {
+      let shouldContinue = this.reading && this.push(this.data.shift());
+      if (!shouldContinue) {
+        this.reading = false;
+      }
+    });
+  }
+}
+
 export function getDockerLogs(name, sessions, args) {
-  const command = 'sudo docker ' + args.join(' ') + ' ' + name;
+  const command = 'sudo docker ' + args.join(' ') + ' ' + name + ' 2>&1';
 
   log(`getDockerLogs command: ${command}`);
 
-  var promises = _.map(sessions, session => {
-    var host = '[' + session._host + ']';
-    var options = {
+  let promises = _.map(sessions, session => {
+    const input = new Callback2Stream();
+    const host = '[' + session._host + ']';
+    const lineSeperator = readline.createInterface({
+      input,
+      terminal: true
+    });
+    lineSeperator.on('line', data => {
+      console.log(host + data);
+    });
+    const options = {
       onStdout: data => {
-        process.stdout.write(host + data);
+        input.addData(data);
       },
       onStderr: data => {
+        // the logs all come in on stdout so stderr isn't added to lineSeperator
         process.stdout.write(host + data);
       }
     };
@@ -57,7 +119,7 @@ export function runSSHCommand(info, command) {
     var sshAgent = process.env.SSH_AUTH_SOCK;
     var ssh = {
       host: info.host,
-      port: info.opts && info.opts.port || 22,
+      port: (info.opts && info.opts.port) || 22,
       username: info.username
     };
 
@@ -78,7 +140,7 @@ export function runSSHCommand(info, command) {
 
     // TODO handle error events
     conn.once('ready', function() {
-      conn.exec(command, function(err, stream) {
+      conn.exec(command, function(err, outputStream) {
         if (err) {
           conn.end();
           reject(err);
@@ -87,11 +149,11 @@ export function runSSHCommand(info, command) {
 
         let output = '';
 
-        stream.on('data', function(data) {
+        outputStream.on('data', function(data) {
           output += data;
         });
 
-        stream.once('close', function(code) {
+        outputStream.once('close', function(code) {
           conn.end();
           resolve({ code, output });
         });